fix(lint): resolve all ruff lint errors across backend
- Sort __all__ exports (RUF022) in 22 __init__.py files
- Prefix unused unpacked variables with underscore (RUF059)
- Remove useless if-else conditions in enforcer.py (RUF034)
- Add stacklevel=2 to warnings.warn calls (B028)
- Remove redundant int() casts (RUF046)
- Parenthesize and/or chains in tests (RUF021)
- Fix Python 3.11 f-string quote reuse in visualizations.py

+++ backend/app/ml/ab_power_analysis.py
@@ -177,13 +177,13 @@ class ABPowerAnalyzer:
             )
 
         # Round up
-        sample_size = int(math.ceil(sample_size))
+        sample_size = math.ceil(sample_size)
 
         # Calculate days to significance
         days = None
         if traffic_per_day and traffic_per_day > 0:
             total_samples = sample_size * 2
-            days = int(math.ceil(total_samples / traffic_per_day))
+            days = math.ceil(total_samples / traffic_per_day)
 
         return PowerAnalysisResult(
             sample_size_per_variant=sample_size,
